image-processor-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3
pillow==11.1.0
pydantic-settings==2.7.1
easyocr==1.7.2
rembg[cpu]==2.0.61
pika==1.3.2